+++ src/atmMachine/ATMCardDetails.java
@@ -2,14 +2,14 @@ package atmMachine;
 
 public class ATMCardDetails {
     private final Integer pin;
-    private Long accountBalance;
+    private int accountBalance;
 
-    public ATMCardDetails(Integer pin, Long accountBalance) {
+    public ATMCardDetails(Integer pin, int accountBalance) {
         this.pin = pin;
         this.accountBalance = accountBalance;
     }
 
-    public Long getAccountBalance() {
+    public int getAccountBalance() {
         return accountBalance;
     }
 
@@ -17,7 +17,7 @@ public class ATMCardDetails {
         return pin;
     }
 
-    public void setAccountBalance(Long accountBalance) {
+    public void setAccountBalance(int accountBalance) {
         this.accountBalance = accountBalance;
     }
 }
